Refactor: Position download folder button in a better parent
