+++ src/frontend/src/pages/MainPage/components/myCollectionComponent/components/headerTabsSearchComponent/index.tsx
@@ -2,16 +2,38 @@ import { useState } from "react";
 import useFlowsManagerStore from "../../../../../../stores/flowsManagerStore";
 import InputSearchComponent from "../inputSearchComponent";
 import TabsSearchComponent from "../tabsComponent";
+import { Button } from "../../../../../../components/ui/button";
+import ForwardedIconComponent from "../../../../../../components/genericIconComponent";
+import useAlertStore from "../../../../../../stores/alertStore";
+import { handleDownloadFolderFn } from "../../../../utils/handle-download-folder";
+import { useFolderStore } from "../../../../../../stores/foldersStore";
+import { useLocation } from "react-router-dom";
 
 type HeaderTabsSearchComponentProps = {};
 
 const HeaderTabsSearchComponent = ({}: HeaderTabsSearchComponentProps) => {
+  const location = useLocation();
+  const myCollectionId = useFolderStore((state) => state.myCollectionId);
+  const folderId = location?.state?.folderId || myCollectionId;
   const isLoading = useFlowsManagerStore((state) => state.isLoading);
   const [tabActive, setTabActive] = useState("Flows");
+  const setErrorData = useAlertStore((state) => state.setErrorData);
+  const allFlows = useFlowsManagerStore((state) => state.allFlows);
+
+  const handleDownloadFolder = () => {
+    if (allFlows.length === 0) {
+      setErrorData({
+        title: "Folder is empty",
+        list: [],
+      });
+      return;
+    }
+    handleDownloadFolderFn(folderId);
+  };
 
   return (
     <>
-      <div className="flex items-end gap-4">
+      <div className="relative flex items-end gap-4">
         <InputSearchComponent loading={isLoading} />
 
         <TabsSearchComponent
@@ -20,6 +42,15 @@ const HeaderTabsSearchComponent = ({}: HeaderTabsSearchComponentProps) => {
           loading={isLoading}
           tabActive={tabActive}
         />
+        <div className="absolute bottom-3 right-0  flex justify-center border-b border-border">
+          <Button variant="primary" onClick={handleDownloadFolder}>
+            <ForwardedIconComponent
+              name="Download"
+              className="main-page-nav-button"
+            />
+            Download
+          </Button>
+        </div>
       </div>
     </>
   );
